+++ databox.js
@@ -1,4 +1,4 @@
 exports.catalog       = require('./lib/catalog.js');
 exports.timeseries    = require('./lib/time-series.js');
 exports.keyValue      = require('./lib/key-value.js');
-exports.notifications = require('./lib/notifications.js');
+exports.subscriptions = require('./lib/subscriptions.js');
